@@ -12,26 +12,26 @@
   "dependencies": {
     "axios": "^0.21.1",
     "blueprint-css": "^3.1.1",
-    "date-fns": "^2.16.0",
+    "date-fns": "^2.21.1",
     "epic-spinners": "^1.1.0",
-    "register-service-worker": "^1.7.1",
-    "swipe-listener": "^1.2.0",
-    "vue": "^2.6.12",
-    "vue-router": "^3.4.3"
+    "register-service-worker": "^1.7.2",
+    "swipe-listener": "^1.3.0",
+    "vue": "^3.0.0-beta.1",
+    "vue-router": "^4.0.6"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.12",
     "@vue/cli-plugin-eslint": "~4.5.12",
     "@vue/cli-plugin-pwa": "~4.5.12",
     "@vue/cli-service": "~4.5.12",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
     "@vue/eslint-config-airbnb": "^5.1.0",
     "@vue/eslint-config-prettier": "6.0.0",
     "eslint": "6.8.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
     "prettier": "^1.19.1",
-    "s3-folder-upload": "^2.3.1",
-    "vue-template-compiler": "^2.6.12"
+    "s3-folder-upload": "^2.3.1"
   },
   "browserslist": [
     "> 1%",
