@@ -1,4 +1,4 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 
 
 import App from './App.vue';
@@ -6,9 +6,4 @@ import router from './router';
 
 require('../node_modules/blueprint-css/dist/blueprint.css');
 
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  render: h => h(App),
-}).$mount('#app');
+createApp(App).use(router).mount('#app');
