@@ -1,21 +1,19 @@
-import Vue from 'vue';
-import Router from 'vue-router';
+import {createRouter, createWebHistory} from 'vue-router';
 import Home from './views/Home.vue';
 
-Vue.use(Router);
-
-export default new Router({
-  mode: 'history',
-  routes: [
-    {
-      path: '/',
-      name: 'home',
-      component: Home,
-    },
-    {
-      path: '/matches/:date',
-      name: 'particular',
-      component: Home,
-    },
-  ],
-});
+const routes = [
+  {
+    path: '/',
+    name: 'home',
+    component: Home,
+  },
+  {
+    path: '/matches/:date',
+    name: 'particular',
+    component: Home,
+  },
+];
+export default createRouter({
+  history: createWebHistory(),
+  routes
+})
